ai-agent/requirements.txt
# AI Agent - Python Dependencies
python-dotenv>=1.0.0
livekit-agents>=0.10.0
livekit-api>=0.8.0
livekit>=1.1.0

# Plugins
livekit-plugins-google>=0.5.0
livekit-plugins-deepgram>=0.5.0
livekit-plugins-silero>=0.5.0
livekit-plugins-elevenlabs>=0.5.0

# Core
numpy>=2.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
